Add a .cargo/config file for binary overrides of tcod-sys building
NOTE: none of this is tested yet.

The general idea is something like this:

1. tcod-sys/build.rs will attept to find libtcod installed in the
well-known system locations (pkg-config and so on).

2. If it fails, it'll try to build it from the attached source

And that's it.

If the user wants to download the precompiled binaries, they can put
them under `precompiled/$target` and uncomment the corresponding section
of `.cargo/config`.

This is the standard way of providing binaries (and it's completely
user-customisable), so we should direct them towards it instead of
bundling the precompiled libs with tcod-sys.

That said, we might still consider doing something like fetching the
binary in the build script if all else fails and it's available.